Reorganize disruptors packages
